[+] Working new TimeInterval

+++ app/model/Event.java
@@ -130,7 +130,7 @@ public class Event extends Model {
 
     public void check() {
         // TODO: verify that event has not already been detected for the last TimeInterval
-        int[] todayTimeInterval = this.getTimeInterval().getActualTimeInterval();
+        long[] todayTimeInterval = this.getTimeInterval().getActualTimeInterval();
         if (EventOccurrence.find.where().eq("event_id", getId())
                 .between("timestamp", todayTimeInterval[0], todayTimeInterval[1])
                 .findIds().size() > 0) { // Event already in DB
